+++ plugins/source/azure/resources/services/cdn/custom_domains.go
@@ -12,8 +12,9 @@ import (
 
 func customDomains() *schema.Table {
 	return &schema.Table{
-		Name:     "azure_cdn_custom_domains",
-		Resolver: fetchCDNCustomDomains,
+		Name:        "azure_cdn_custom_domains",
+		Description: `https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/services/cdn/mgmt/2020-09-01/cdn#CustomDomain`,
+		Resolver:    fetchCDNCustomDomains,
 		Columns: []schema.Column{
 			{
 				Name:     "subscription_id",