+++ plugins/source/azure/resources/services/sql/encryption_protectors.go
@@ -13,8 +13,9 @@ import (
 
 func encryptionProtectors() *schema.Table {
 	return &schema.Table{
-		Name:     "azure_sql_encryption_protectors",
-		Resolver: fetchSQLEncryptionProtectors,
+		Name:        "azure_sql_encryption_protectors",
+		Description: `https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/services/preview/sql/mgmt/v4.0/sql#EncryptionProtector`,
+		Resolver:    fetchSQLEncryptionProtectors,
 		Columns: []schema.Column{
 			{
 				Name:     "subscription_id",