+++ plugins/source/azure/resources/services/sql/server_security_alert_policies.go
@@ -13,8 +13,9 @@ import (
 
 func serverSecurityAlertPolicies() *schema.Table {
 	return &schema.Table{
-		Name:     "azure_sql_server_security_alert_policies",
-		Resolver: fetchSQLServerSecurityAlertPolicies,
+		Name:        "azure_sql_server_security_alert_policies",
+		Description: `https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/services/preview/sql/mgmt/v4.0/sql#ServerSecurityAlertPolicy`,
+		Resolver:    fetchSQLServerSecurityAlertPolicies,
 		Columns: []schema.Column{
 			{
 				Name:     "subscription_id",