+++ plugins/source/heroku/resources/services/team_space.go
@@ -13,7 +13,7 @@ import (
 func TeamSpaces() *schema.Table {
 	return &schema.Table{
 		Name:        "heroku_team_spaces",
-		Description: `https://devcenter.heroku.com/articles/platform-api-reference#team-space-attributes`,
+		Description: `https://devcenter.heroku.com/articles/platform-api-reference#team-space`,
 		Resolver:    fetchTeamSpaces,
 		Columns: []schema.Column{
 			{